feat(zentrain): canonical-dataset tooling — pareto converter, effort-cap degradation fitter, multi-arm trainer runner
- zentrain/tools/canonical_to_pareto.py: canonical 2026-06-27 per-codec parquet →
  the PARETO parquet + FEATURES TSV train_hybrid.py consumes (config_name = cell,
  size_class re-derived from width×height — the canonical column is 'large' on
  every row — named feat_* only, one verified-constant features row per image).
  The 2026-07-03 zenjxl-modular bake did this in an uncommitted scratch script.
- zentrain/examples/canonical_picker_config.py: one categorical-cell config for
  every codec (ZENCANONICAL_CODEC / ZENCANONICAL_DIR).
- zentrain/tools/fit_family_degradation.py: per-family effort-cap degradation
  (ln bytes_at(cap) − ln bytes_at(ref) per image × target, per size class, with
  reach) + per-effort encode cost (ms/MP, ln ms = a + b·ln px); knob semantics per
  codec (webp m / jxl e ascending, rav1e s inverted). Tables committed for
  zenwebp (train+validate+test), zenjxl and zenavif (validate):
  benchmarks/family_degradation_<codec>_2026-08-28.{tsv,md}.
- zentrain/tools/leakyrelu_seeds_runner.py rewritten as a general multi-arm
  runner (--arms relu,leakyrelu,leakyrelu+wd=1e-4, --reuse, --report); the old
  one hard-coded /home/lilith zenjpeg paths and a train_hybrid location that no
  longer exists.
- fix(train_hybrid): origin even/odd split built tr/va/te with np.array([...]);
  an empty bucket was float64 and Xe[te] raised 'arrays used as indices must be
  of integer (or boolean) type'. dtype pinned to int64 (seen on the canonical
  zenwebp train+validate pair).
- 7 tests in zentrain/tools/test_canonical_tools.py, added to the CI
  zentrain-pytests job and the justfile recipe.

refs #68 refs #85

+++ zentrain/tools/train_hybrid.py
@@ -3547,9 +3547,14 @@ def main():
         )
     _spl = [_origin_split_of(m[0]) for m in meta]
     n_unsplit = sum(1 for s in _spl if s is None)
-    tr = np.array([i for i, s in enumerate(_spl) if s == "train"])
-    va = np.array([i for i, s in enumerate(_spl) if s == "val"])
-    te = np.array([i for i, s in enumerate(_spl) if s == "test"])
+    # dtype pinned: an EMPTY bucket (e.g. a corpus with no 7/9-digit origins, or
+    # train+validate canonical splits fed without test) would otherwise be a
+    # float64 array and `Xe[te]` raises "arrays used as indices must be of
+    # integer (or boolean) type" — seen 2026-08-28 on the canonical zenwebp
+    # train+validate pair.
+    tr = np.array([i for i, s in enumerate(_spl) if s == "train"], dtype=np.int64)
+    va = np.array([i for i, s in enumerate(_spl) if s == "val"], dtype=np.int64)
+    te = np.array([i for i, s in enumerate(_spl) if s == "test"], dtype=np.int64)
     n_origins = len({_origin_stem(m[0]) for m in meta})
     n_rend = len({m[0] for m in meta})
     sys.stderr.write(
